# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для Marketplace приложения.",
        "title": "Marketplace API",
        "termsOfService": "http://your-website.com/terms/",
        "contact": {
            "name": "Support Team",
            "url": "http://your-website.com/support",
            "email": "support@your-website.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "195.49.215.120:8080",
    "basePath": "/",
    "paths": {
...
        "/api/categories/subcategories": {
            "get": {
                "description": "Возвращает список всех подкатегорий для заданного path категории.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Категории"
                ],
                "summary": "Получение подкатегорий по path",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Path категории",
                        "name": "path",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_WhyDias_Marketplace_internal_controllers.Category"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_WhyDias_Marketplace_internal_controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_WhyDias_Marketplace_internal_controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/products": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Добавляет новый продукт с изображениями и вариациями",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Продукты"
                ],
                "summary": "Добавление нового продукта",
                "parameters": [
                    {
                        "description": "Данные продукта",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_WhyDias_Marketplace_internal_controllers.AddProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Продукт успешно добавлен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/products/moderated": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список продуктов, находящихся на модерации (status_id = 3).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Продукты"
                ],
                "summary": "Получение продуктов с модерацией",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_WhyDias_Marketplace_internal_models.Product"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_WhyDias_Marketplace_internal_controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_WhyDias_Marketplace_internal_controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/products/unmoderated": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список продуктов, прошедших модерацию (status_id = 2).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Продукты"
                ],
                "summary": "Получение продуктов без модерации",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_WhyDias_Marketplace_internal_models.Product"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_WhyDias_Marketplace_internal_controllers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_WhyDias_Marketplace_internal_controllers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_WhyDias_Marketplace_internal_controllers.AddProductRequest": {
            "type": "object",
            "required": [
                "category_id",
                "name"
            ],
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "variations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_WhyDias_Marketplace_internal_controllers.ProductVariationRequest"
                    }
                }
            }
        },
        "github_com_WhyDias_Marketplace_internal_controllers.AttributeValueRequest": {
            "type": "object",
            "required": [
                "name",
                "value"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "github_com_WhyDias_Marketplace_internal_controllers.Category": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
        "github_com_WhyDias_Marketplace_internal_controllers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_WhyDias_Marketplace_internal_controllers.ProductVariationRequest": {
            "type": "object",
            "required": [
                "price",
                "sku",
                "stock"
            ],
            "properties": {
                "attributes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_WhyDias_Marketplace_internal_controllers.AttributeValueRequest"
                    }
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "price": {
                    "type": "number"
                },
                "sku": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_WhyDias_Marketplace_internal_models.AttributeValue": {
            "type": "object",
            "properties": {
                "attribute_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "github_com_WhyDias_Marketplace_internal_models.Product": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_WhyDias_Marketplace_internal_models.ProductImage"
                    }
                },
                "market_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status_id": {
                    "type": "integer"
                },
                "supplier_id": {
                    "type": "integer"
                },
                "variations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_WhyDias_Marketplace_internal_models.ProductVariation"
                    }
                }
            }
        },
        "github_com_WhyDias_Marketplace_internal_models.ProductImage": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "image_urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "product_id": {
                    "type": "integer"
                }
            }
        },
        "github_com_WhyDias_Marketplace_internal_models.ProductVariation": {
            "type": "object",
            "properties": {
                "attributes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_WhyDias_Marketplace_internal_models.AttributeValue"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_WhyDias_Marketplace_internal_models.ProductVariationImage"
                    }
                },
                "price": {
                    "type": "number"
                },
                "product_id": {
                    "type": "integer"
                },
                "sku": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                }
            }
        },
        "github_com_WhyDias_Marketplace_internal_models.ProductVariationImage": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "image_urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "product_variation_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}